requirements.txt
lxml
discord.py
